pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "multi-bot-ai-system"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111,<0.112",
    "uvicorn[standard]>=0.30,<0.31",
    "pydantic>=2.7,<3",
    "httpx>=0.27,<0.28",
    "redis[asyncio]>=5.0,<6",
    "asyncpg>=0.29,<0.30",
    "python-telegram-bot>=21.0,<22",
    "python-jose[cryptography]>=3.3,<4",
    "apscheduler>=3.10,<4",
    "feedparser>=6.0,<7",
    "python-dotenv>=1.0,<2",
]

[project.optional-dependencies]
test = [
    "hypothesis>=6.100,<7",
    "pytest>=8.0,<9",
    "pytest-asyncio>=0.23,<0.24",
    "pytest-mock>=3.14,<4",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*"]
